VALE SHANE WARNE
Shane Keith Warne (13 September 1969 – 4 March 2022) was an Australian cricketer. A right-arm leg spinner, he is widely considered as one of the greatest bowlers in cricket history, and in 2000 he was selected by a panel of cricket experts as one of five Wisden Cricketers of the Century, the only specialist bowler and the only one still playing at the time.